Expose generate via Graphula.Arbitrary
Anything in tests that utilizes random generation can use the graphula
seed and remain deterministic, instead of falling back to `IO`.
